refactor sidebar: icon are now icon-comp. animation trigger only after first human interaction. state of sidebar (shrink, expand) is stored. icon source is now a CDN, local file is backup
