--- src/app/ui/shared/components/atoms/icon/icon.component.scss
@@ -1,3 +1,0 @@
-.app-icon {
-  display: inline-block;
-}
