+++ src/app/ui/shared/components/atoms/icon/icon.component.ts
@@ -1,11 +1,10 @@
 import { Component, input } from '@angular/core';
-import { PossibleIconNames } from './icon.model';
+import { PossibleIconNames } from '../../../model/icons/icon.model';
 
 @Component({
   selector: 'app-icon',
   imports: [],
   templateUrl: './icon.component.html',
-  styleUrl: './icon.component.scss',
   host: {
     role: 'img',
     class: 'app-icon',
