+++ src/app/ui/shared/components/layout/sidebar/sidebar.component.ts
@@ -8,12 +8,13 @@ import {
 import { SidebarItem } from './sidebar.model';
 import { RouterLink, RouterLinkActive } from '@angular/router';
 import { environment } from '@src/environments/environment';
+import { IconComponent } from '../../atoms/icon/icon.component';
 
 const version = environment.version;
 
 @Component({
   selector: 'app-sidebar',
-  imports: [RouterLink, RouterLinkActive],
+  imports: [RouterLink, RouterLinkActive, IconComponent],
   templateUrl: './sidebar.component.html',
   styleUrl: './sidebar.component.scss',
   encapsulation: ViewEncapsulation.None
@@ -28,9 +29,19 @@ export class SidebarComponent implements OnInit {
   expanded = input<boolean>(true);
 
   private readonly bodyTag = signal<HTMLBodyElement | null>(null);
-  public expandedStatus = signal<boolean>(this.expanded());
+  public expandedStatus = signal<boolean>(this.getInitialExpandedStatus());
   public appVersion = signal<string>('');
   public appLogoRoot = signal<string>('in-app-logo/');
+
+  private getInitialExpandedStatus(): boolean {
+    const storedSidebarStatus = localStorage.getItem('expandedStatus');
+
+    if (storedSidebarStatus !== null) {
+      return JSON.parse(storedSidebarStatus);
+    }
+
+    return this.expanded();
+  }
 
   ngOnInit(): void {
     this.bodyTag.set(document.querySelector('body'));
@@ -38,21 +49,33 @@ export class SidebarComponent implements OnInit {
     if (this.bodyTag()) {
       this.bodyTag()!.classList.add('has-sidebar');
 
-      if (this.expandedStatus() === true) {
+      if (this.expandedStatus()) {
         this.bodyTag()!.classList.add('has-sidebar--expanded');
       }
     }
+
+    localStorage.setItem(
+      'expandedStatus',
+      JSON.stringify(this.expandedStatus())
+    );
   }
 
   toggleMenu(): void {
     this.expandedStatus.set(!this.expandedStatus());
 
+    localStorage.setItem(
+      'expandedStatus',
+      JSON.stringify(this.expandedStatus())
+    );
+
     if (this.bodyTag()) {
-      if (this.expandedStatus() === true) {
+      if (this.expandedStatus()) {
         this.bodyTag()!.classList.add('has-sidebar--expanded');
-      } else if (this.expandedStatus() === false) {
+      } else {
         this.bodyTag()!.classList.remove('has-sidebar--expanded');
       }
+
+      this.bodyTag()!.classList.add('has-sidebar--animation');
     }
   }
