+++ src/app/ui/shared/components/layout/sidebar/sidebar.model.ts
@@ -1,6 +1,8 @@
+import { PossibleIconNames } from '../../../model/icons/icon.model';
+
 export interface SidebarItem {
   id: string;
-  icon: string;
+  icon: PossibleIconNames;
   label: string;
   shortLabel?: string;
   route: string;
